+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/BuildSearchableOptionsTask.kt
@@ -13,12 +13,12 @@ import org.jetbrains.intellij.platform.gradle.Constants.Tasks
 import org.jetbrains.intellij.platform.gradle.GradleProperties
 import org.jetbrains.intellij.platform.gradle.extensions.IntelliJPlatformExtension
 import org.jetbrains.intellij.platform.gradle.get
+import org.jetbrains.intellij.platform.gradle.services.PluginXmlService
+import org.jetbrains.intellij.platform.gradle.services.pluginXmlService
 import org.jetbrains.intellij.platform.gradle.tasks.RunIdeTask.Companion.systemPropertyDefault
 import org.jetbrains.intellij.platform.gradle.tasks.aware.RunnableIdeAware
-import org.jetbrains.intellij.platform.gradle.tasks.aware.parse
 import org.jetbrains.intellij.platform.gradle.utils.Logger
 import org.jetbrains.intellij.platform.gradle.utils.asPath
-import org.jetbrains.intellij.platform.gradle.utils.extensionProvider
 import org.jetbrains.intellij.platform.gradle.utils.safePathString
 
 /**
@@ -26,6 +26,8 @@ import org.jetbrains.intellij.platform.gradle.utils.safePathString
  * This task runs a headless IDE instance to collect all the available options provided by the plugin's [Settings](https://plugins.jetbrains.com/docs/intellij/settings.html).
  *
  * If the plugin doesn't implement custom settings, it is recommended to disable this task via [IntelliJPlatformExtension.buildSearchableOptions] flag.
+ * The task is also skipped automatically when no Configurable extension points are found in the main plugin descriptor
+ * or plugin module descriptors. Use [GradleProperties.ForceBuildSearchableOptions] to bypass this optimization.
  *
  * In the case of running the task for the plugin that has [IntelliJPlatformExtension.PluginConfiguration.ProductDescriptor] defined,
  * a warning will be logged regarding potential issues with running headless IDE for paid plugins.
@@ -48,6 +50,9 @@ abstract class BuildSearchableOptionsTask : JavaExec(), RunnableIdeAware {
      */
     @get:Internal
     abstract val showPaidPluginWarning: Property<Boolean>
+
+    @get:Internal
+    abstract val pluginXmlService: Property<PluginXmlService>
 
     private val log = Logger(javaClass)
 
@@ -83,8 +88,10 @@ abstract class BuildSearchableOptionsTask : JavaExec(), RunnableIdeAware {
     companion object : Registrable {
         override fun register(project: Project) =
             project.registerTask<BuildSearchableOptionsTask>(Tasks.BUILD_SEARCHABLE_OPTIONS) {
-                val buildSearchableOptionsEnabledProvider = project.extensionProvider.flatMap { it.buildSearchableOptions }
+                val buildSearchableOptionsEnabledProvider = project.buildSearchableOptionsEnabledProvider()
                 val prepareSandboxTaskProvider = project.tasks.named<PrepareSandboxTask>(Tasks.PREPARE_SANDBOX)
+                val pluginXmlServiceProvider = project.pluginXmlService()
+                val pluginXmlProvider = pluginXml
                 applySandboxFrom(prepareSandboxTaskProvider)
 
                 outputDirectory.convention(
@@ -92,15 +99,17 @@ abstract class BuildSearchableOptionsTask : JavaExec(), RunnableIdeAware {
                         temporaryDir
                     })
                 )
+                pluginXmlService.convention(pluginXmlServiceProvider)
                 showPaidPluginWarning.convention(
-                    project.providers[GradleProperties.PaidPluginSearchableOptionsWarning].map {
-                        it && pluginXml.orNull?.parse { productDescriptor } != null
+                    project.providers[GradleProperties.PaidPluginSearchableOptionsWarning].map { enabled ->
+                        enabled && pluginXmlProvider.orNull?.let { pluginXmlServiceProvider.get().resolve(it.asPath).productDescriptor } != null
                     }
                 )
 
                 systemPropertyDefault("idea.l10n.keys", "only")
 
                 inputs.property("intellijPlatform.buildSearchableOptions", buildSearchableOptionsEnabledProvider)
+                inputs.property("intellijPlatform.forceBuildSearchableOptions", project.providers[GradleProperties.ForceBuildSearchableOptions])
 
                 onlyIf {
                     buildSearchableOptionsEnabledProvider.get()