+++ src/test/kotlin/org/jetbrains/intellij/platform/gradle/tasks/InitializeIntelliJPlatformPluginTaskTest.kt
@@ -100,4 +100,44 @@ class InitializeIntelliJPlatformPluginTaskTest : IntelliJPluginTestBase() {
             assertNotContains("${Plugin.NAME} is outdated: 0.0.0.", output)
         }
     }
+
+    @Test
+    fun `module project tasks do not depend on initialize task`() {
+        buildFile overwrite //language=kotlin
+                """
+                import org.jetbrains.intellij.platform.gradle.*
+                import org.jetbrains.intellij.platform.gradle.tasks.*
+                import org.jetbrains.kotlin.gradle.dsl.*
+
+                version = "1.0.0"
+
+                plugins {
+                    id("java")
+                    id("org.jetbrains.intellij.platform.module")
+                    id("org.jetbrains.kotlin.jvm") version "$kotlinPluginVersion"
+                }
+
+                kotlin {
+                    jvmToolchain(21)
+                }
+
+                repositories {
+                    mavenCentral()
+
+                    intellijPlatform {
+                        defaultRepositories()
+                    }
+                }
+
+                dependencies {
+                    intellijPlatform {
+                        create("$intellijPlatformType", "$intellijPlatformVersion")
+                    }
+                }
+                """.trimIndent()
+
+        build(Tasks.GENERATE_MANIFEST) {
+            assertNotContains(":${Tasks.INITIALIZE_INTELLIJ_PLATFORM_PLUGIN}", output)
+        }
+    }
 }